cleanup merge conflicts
